marketplace-ui/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
mangum==0.17.0
snowflake-connector-python
pydantic==2.7.0
python-dotenv==1.0.1
